[Fix] Move CastRemover to ir_util.h
