+++ tvm/src/pass/ir_util.h
@@ -8,6 +8,7 @@
 
 #include <tvm/ir.h>
 #include <tvm/runtime/device_api.h>
+#include <tvm/ir_mutator.h>
 #include <vector>
 
 namespace TVM {
@@ -154,6 +155,124 @@ inline int GetTempAllocaAlignment(Type type, int32_t const_size) {
   }
   return align;
 }
+
+// Remove cast in binary expressions
+// Example: (cast(x) + cast(y)) -> (x + y)
+// Usage: CastRemover castRemover;
+//        Expr expr = castRemover.Mutate(expr);
+class CastRemover final : public IRMutator {
+ public:
+  CastRemover() {}
+
+  Expr Mutate_(const Cast* op, const Expr& e) {
+    return op->value;
+  }
+
+  Expr Mutate_(const Add* op, const Expr& e) {
+    Expr a = this->Mutate(op->a);
+    Expr b = this->Mutate(op->b);
+    if (const Cast* ca = a.as<Cast>()) {
+      a = ca->value;
+    }
+    if (const Cast* cb = b.as<Cast>()) {
+      b = cb->value;
+    }
+    if (a.type() != b.type())
+      LOG(FATAL) << "CastRemover: type mismatch "
+        << a.type() << " vs " << b.type();
+    return Add::make(a, b);
+  }
+
+  Expr Mutate_(const Sub* op, const Expr& e) {
+    Expr a = this->Mutate(op->a);
+    Expr b = this->Mutate(op->b);
+    if (const Cast* ca = a.as<Cast>()) {
+      a = ca->value;
+    }
+    if (const Cast* cb = b.as<Cast>()) {
+      b = cb->value;
+    }
+    if (a.type() != b.type())
+      LOG(FATAL) << "CastRemover: type mismatch "
+        << a.type() << " vs " << b.type();
+    return Sub::make(a, b);
+  }
+
+  Expr Mutate_(const Mul* op, const Expr& e) {
+    Expr a = this->Mutate(op->a);
+    Expr b = this->Mutate(op->b);
+    if (const Cast* ca = a.as<Cast>()) {
+      a = ca->value;
+    }
+    if (const Cast* cb = b.as<Cast>()) {
+      b = cb->value;
+    }
+    if (a.type() != b.type())
+      LOG(FATAL) << "CastRemover: type mismatch "
+        << a.type() << " vs " << b.type();
+    return Mul::make(a, b);
+  }
+
+  Expr Mutate_(const Div* op, const Expr& e) {
+    Expr a = this->Mutate(op->a);
+    Expr b = this->Mutate(op->b);
+    if (const Cast* ca = a.as<Cast>()) {
+      a = ca->value;
+    }
+    if (const Cast* cb = b.as<Cast>()) {
+      b = cb->value;
+    }
+    if (a.type() != b.type())
+      LOG(FATAL) << "CastRemover: type mismatch "
+        << a.type() << " vs " << b.type();
+    return Div::make(a, b);
+  }
+
+  Expr Mutate_(const Mod* op, const Expr& e) {
+    Expr a = this->Mutate(op->a);
+    Expr b = this->Mutate(op->b);
+    if (const Cast* ca = a.as<Cast>()) {
+      a = ca->value;
+    }
+    if (const Cast* cb = b.as<Cast>()) {
+      b = cb->value;
+    }
+    if (a.type() != b.type())
+      LOG(FATAL) << "CastRemover: type mismatch "
+        << a.type() << " vs " << b.type();
+    return Mod::make(a, b);
+  }
+
+  Expr Mutate_(const Min* op, const Expr& e) {
+    Expr a = this->Mutate(op->a);
+    Expr b = this->Mutate(op->b);
+    if (const Cast* ca = a.as<Cast>()) {
+      a = ca->value;
+    }
+    if (const Cast* cb = b.as<Cast>()) {
+      b = cb->value;
+    }
+    if (a.type() != b.type())
+      LOG(FATAL) << "CastRemover: type mismatch "
+        << a.type() << " vs " << b.type();
+    return Min::make(a, b);
+  }
+
+  Expr Mutate_(const Max* op, const Expr& e) {
+    Expr a = this->Mutate(op->a);
+    Expr b = this->Mutate(op->b);
+    if (const Cast* ca = a.as<Cast>()) {
+      a = ca->value;
+    }
+    if (const Cast* cb = b.as<Cast>()) {
+      b = cb->value;
+    }
+    if (a.type() != b.type())
+      LOG(FATAL) << "CastRemover: type mismatch "
+        << a.type() << " vs " << b.type();
+    return Max::make(a, b);
+  }
+};
 }  // namespace ir
 }  // namespace TVM
 #endif  // PASS_IR_UTIL_H_
